Update locales-nl-NL.xml
In Chicago 17th (full note) in the Dutch locale, single quotation marks are returned, but they should be double. This issue is both in the notes and the bibliography.

Example: John H. Hanson, ‘Jihad and the Ahmadiyya Muslim Community: Nonviolent Efforts to Promote Islam in the Contemporary World’, Nova Religio: The Journal of Alternative and Emergent Religions 11, nr. 2 (2007): 85, https://doi.org/10.1525/nr.2007.11.2.77.